rear_end/requirements.txt
Django>=6.0.2,<7.0
mysqlclient>=2.2.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
faiss-cpu>=1.8.0
